requirements.txt
mypy
pylint
PyYAML
pandas
requests
openpyxl
python-dotenv
psycopg2-binary
sqlalchemy
